# backend/openapi.yaml
openapi: 3.0.3
info:
  title: EduVerify API
  version: "1.0.0"
  description: >
    API REST del backend de EduVerify. Todas las respuestas siguen el envelope
    { status: 'success', data } / { status: 'error', message }.
    Ver backend/README.md para setup y arquitectura.
servers:
  - url: http://localhost:3001
security: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Login con email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [correo, password]
              properties:
                correo: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "400": { $ref: '#/components/responses/BadRequest' }
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/auth/registro:
    post:
      tags: [auth]
      summary: Crear cuenta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, correo, password]
              properties:
                nombre: { type: string, minLength: 2, maxLength: 120 }
                correo: { type: string, format: email }
                password: { type: string, minLength: 8 }
                rol:
                  type: string
                  enum: [estudiante, profesor, creador]
                  default: estudiante
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "409":
          description: Email ya registrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/auth/google:
    post:
      tags: [auth]
      summary: Login/registro vía Google Identity Services
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credential]
              properties:
                credential:
                  type: string
                  description: JWT credential emitido por Google GSI
      responses:
        "200":
          description: Login exitoso (crea la cuenta si no existía)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
  /api/auth/cambiar-password:
    post:
      tags: [auth]
      summary: Solicitar email de recuperación de contraseña
      description: Siempre responde 200 exista o no el email, para no filtrar qué cuentas existen.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Email enviado si la cuenta existe
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "400": { $ref: '#/components/responses/BadRequest' }
  /api/auth/actualizar-password:
    post:
      tags: [auth]
      summary: Confirmar reseteo de contraseña con token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, token, password]
              properties:
                id: { type: integer, description: user id (viene del link del email) }
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Contraseña actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "400":
          description: Token inválido/expirado o body inválido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ── USERS ─────────────────────────────────────────────────────────────────
...
  /api/videos/{videoId}/comments:
    get:
      tags: [comments]
      summary: Listar comentarios de un video (anidados por parent_id)
      description: >
        Si se envía Authorization, cada comentario incluye `liked` según los likes
        del usuario autenticado; sin token, `liked` es siempre false.
      security: [{ bearerAuth: [] }, {}]
      parameters:
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Comment' }
    post:
      tags: [comments]
      summary: Crear comentario o respuesta
      security: [{ bearerAuth: [] }]
      parameters:
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [texto]
              properties:
                texto: { type: string, minLength: 1, maxLength: 2000 }
                parent_id:
                  type: integer
                  nullable: true
                  description: id del comentario padre para responder
      responses:
        "201":
          description: Comentario creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Comment' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/comments/{id}/like:
    post:
      tags: [comments]
      summary: Toggle like de un comentario
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      likes: { type: integer }
                      liked: { type: boolean }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/comments/{id}:
    delete:
      tags: [comments]
      summary: Eliminar comentario (solo el autor)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403":
          description: No es el autor del comentario
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── FAVORITES ─────────────────────────────────────────────────────────────
...
  /api/history:
    get:
      tags: [history]
      summary: Historial de reproducción (ordenado por más reciente)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Video' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [history]
      summary: Vaciar todo el historial
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/history/{videoId}:
    post:
      tags: [history]
      summary: Registrar reproducción de un video (upsert watched_at)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [history]
      summary: Eliminar una entrada del historial
      security: [{ bearerAuth: [] }]
      parameters:
        - name: videoId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
        "401": { $ref: '#/components/responses/Unauthorized' }

  # ── PLAYLISTS (estudiante) ───────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token obtenido de /api/auth/login, /registro o /google. Payload:
        { sub: userId, email, rol, premium }. Expira en 7 días.
...
  responses:
    BadRequest:
      description: Body/params inválidos (Zod)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Token ausente, inválido o expirado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Categoria:
      type: string
      enum: [Programación, Ciberseguridad, Matemáticas, Electrónica, Arte]
    ErrorResponse:
      type: object
      properties:
        status: { type: string, example: error }
        message: { type: string }
    SuccessMessage:
      type: object
      properties:
        status: { type: string, example: success }
        data:
          type: object
          properties:
            message: { type: string }
    User:
      type: object
      properties:
        id: { type: integer }
        nombre: { type: string }
        email: { type: string, format: email }
        rol: { type: string, enum: [estudiante, profesor, creador] }
        premium: { type: boolean }
        dark_mode: { type: boolean }
        avatar_url: { type: string, format: uri, nullable: true }
        banner_url: { type: string, format: uri, nullable: true }
    AuthResponse:
      type: object
      properties:
        status: { type: string, example: success }
        data:
          type: object
          properties:
            token: { type: string }
            user: { $ref: '#/components/schemas/User' }
    Video:
      type: object
      properties:
        id: { type: integer }
        titulo: { type: string }
        descripcion: { type: string, nullable: true }
        url_video: { type: string, format: uri }
        categoria: { $ref: '#/components/schemas/Categoria' }
        tipo: { type: string, enum: [grabado, envivo] }
        es_premium: { type: boolean }
        vistas: { type: integer }
        duracion: { type: string }
        created_at: { type: string, format: date-time }
        autor: { type: string }
        autor_id: { type: integer }
        author_avatar_url: { type: string, format: uri, nullable: true }
...
    Comment:
      type: object
      properties:
        id: { type: integer }
        video_id: { type: integer }
        user_id: { type: integer }
        parent_id: { type: integer, nullable: true }
        texto: { type: string }
        likes: { type: integer }
        created_at: { type: string, format: date-time }
        autor: { type: string }
        rol: { type: string, enum: [estudiante, profesor, creador] }
        autor_avatar_url: { type: string, format: uri, nullable: true }
        liked: { type: boolean }
        respuestas:
          type: array
          items: { $ref: '#/components/schemas/Comment' }
...